refactor: use new hook instead of useWindowEvents
